requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==3.4.0
pandas==1.5.2
pandas-gbq==0.18.1

